refactor(ios): extract native glass components
